refactor(worker): rename jobs/ to kyc/, rename JOB_NAMES to INDEXER_JOB_NAMES (#5)
- jobs/ was a generic name for a single KYC concern — renamed to kyc/
- JOB_NAMES was ambiguous — renamed to INDEXER_JOB_NAMES since it only
  controls indexer job discovery, not the KYC outbox job
