+++ .env.example
@@ -2,7 +2,7 @@
 POSTGRES_URL=postgres://user:password@localhost:5432/pipeline
 
 # Comma-separated list of job names (e.g. ETH_MAINNET)
-JOB_NAMES=
+INDEXER_JOB_NAMES=
 
 # Per-job config — replace <NAME> with the uppercased job name (e.g. ETH_MAINNET)
 JOB_<NAME>_ENABLED=true
