+++ packages/worker/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod config;
 pub mod indexer;
-pub mod jobs;
+pub mod kyc;
